refactor to clean up moduels
